ref(settings): move settings files under their own directory
