+++ rootfs/manage.py
@@ -5,7 +5,7 @@ import sys
 
 
 if __name__ == "__main__":
-    os.environ.setdefault("DJANGO_SETTINGS_MODULE", "deis.settings")
+    os.environ.setdefault("DJANGO_SETTINGS_MODULE", "deis.settings.production")
 
     from django.core.management import execute_from_command_line
